+++ Changelog.md
@@ -1,6 +1,7 @@
 TenshiLib 1.6.14
 ================
 - Fix patreon data getter not being async
+- Add a color slider and alpha support for patreon effects
 
 TenshiLib 1.6.13
 ================